+++ src/__tests__/utils/storage.test.ts
@@ -1,4 +1,4 @@
-import { storage } from '../../utils/storage';
+import { storage } from '../../echo-base/utils/storage';
 
 describe('Storage', () => {
     function mockLocalStorage<T>(mockObj: T): T {
