+++ src/plants/persistPlantsData.ts
@@ -1,7 +1,7 @@
+import { storage } from '../echo-base/utils/storage';
 import { plantsData } from '../state/defaultStates';
 import { PlantsData } from '../types/plants';
 import { EchoLocalStorage } from '../types/storage';
-import { storage } from '../utils/storage';
 
 export class PersistPlantsData {
     private echoStorage: EchoLocalStorage;
